+++ src/core/data/middlewares/hydration-cache/CHANGELOG.md
@@ -11,6 +11,10 @@ Changelog
 
 ## v4.0.0-beta.?? (2024-??-??)
 
+#### :boom: Breaking Change
+
+* Added a wrapper for middleware with additional parameters
+
 #### :rocket: New Feature
 
 * Initial release